Better collocate XCTFail source code context (#67)
XCTest usually does a pretty good job of attaching an out-of-test
XCTFail to the source code context that makes it most visible to Xcode
users, but for complex tests it seems to stop working after awhile, and
it also seems to have gotten less reliable in Xcode 15.

So we can do our best to take over. We can use @tgrapperon's call stack
code to find the test frame and manually assign it to the issue's source
code context.